implemented function tracing
implemented a new session method .at_function(...) wherein the
user can specify the address of a function and prototype its
arguments. when that function is executed, mozz will invoke the
callback function with all the function arguments pulled from
the appropriate registers/stack positions and converted to python
values.

a "Convention" class must exist for the calling convention of the
binary for this feature to work. right now x86_64 sysv calling
convention is the only implemented convention.